+++ src/main/java/com/product/review/reviewservice/service/ReviewService.java
@@ -7,7 +7,7 @@ import java.util.Optional;
 public interface ReviewService {
 
     public Review addProdcutReview(Review review);
-    public Iterable<Review>  findProductReviews(int product_id);
+    public Iterable<Review>  findProductReviews(Long product_id);
     public Review updateproductReview(Review review);
     public Optional<Review> findProductReview(Long id) ;
