+++ src/main/java/com/product/review/reviewservice/service/ReviewServiceImpl.java
@@ -18,7 +18,7 @@ public class ReviewServiceImpl implements ReviewService {
     }
 
     @Override
-    public Iterable<Review> findProductReviews(int product_id) {
+    public Iterable<Review> findProductReviews(Long product_id) {
 
         return reviewServiceDao.findByProductId(product_id);
     }
